+++ aplicativoEcommerce/EcommerceAPI/app/Http/Controllers/ItemPedidoController.php
@@ -29,21 +29,29 @@ class ItemPedidoController extends Controller
 
     public function show($id)
     {
-        $itemPedido = ItemPedido::findOrFail($id);
+        $itemPedido = ItemPedido::find($id);
+        if (!$itemPedido) {
+            return response()->json(['mensagem' => 'Item Pedido não encontrado'], 404);
+        }
         return response()->json($itemPedido);
     }
 
     public function destroy($id)
     {
-        $itemPedido = ItemPedido::findOrFail($id);
+        $itemPedido = ItemPedido::find($id);
+        if (!$itemPedido) {
+            return response()->json(['mensagem' => 'Item Pedido não encontrado'], 404);
+        }
         $itemPedido->delete();
     return response()->json(["mensagem" => "ItemPedido removido com sucesso!"]);
     }
 
     public function update(Request $request, $id)
     {
-        $itemPedido = ItemPedido::findOrFail($id);
-
+        $itemPedido = ItemPedido::find($id);
+        if (!$itemPedido) {
+            return response()->json(['mensagem' => 'Item Pedido não encontrado'], 404);
+        }
         $validated = $request->validate([
             'pedido_id' => 'required|integer|exists:pedidos,id',
             'produto_id' => 'required|integer|exists:produtos,id',
